requirements.txt
attrs==21.2.0
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
elasticsearch==7.12.0
fake-useragent==0.1.11
Faker==8.8.1
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.18.0
queuelib==1.6.1
requests==2.25.1
schedule==1.1.0
Scrapy==2.5.0
scrapy-fake-useragent==1.4.4
service-identity==21.1.0
six==1.16.0
text-unidecode==1.3
Twisted==21.2.0
urllib3==1.26.5
w3lib==1.22.0
zope.interface==5.4.0